webapp/terminal guide: some refactor
